CoinTracker.ino
This project will help the people to track crytpo coins better.
This is the code of our project!!
You will have a switch to control your coin data. Every side of your switch tracks 2 diffrent coins in the same time. When you change the condition, it starts to track other 2 coins. That's why you are having a better chance to control your coins with one 16x2 LCD via this project